Added random colours to player bullets plus Impact shatter effect and cleaned key location clear sphere script in process
