+++ Assets/Prefabs/KeyLocation.prefab
@@ -49,9 +49,11 @@ MonoBehaviour:
   radius: 0
   sphere: {fileID: 3549167572352077796}
   clearSphere: {fileID: 7324278031975958964, guid: ba56b467c4ceafc46a377431fd8d74d0, type: 3}
-  clearSphereSpawnAmount: 15
+  clearSphereSpawnAmount: 30
   players: []
-  speedMod: 0
+  fountainWaterObject: {fileID: 0}
+  cleanFountainMain: {r: 0, g: 0, b: 0, a: 0}
+  cleanFountainSecondary: {r: 0, g: 0, b: 0, a: 0}
   clearedClip: {fileID: 8300000, guid: a5f9e6ddbd472fb408b10a32c9d9783b, type: 3}
 --- !u!1 &3549167572352077796
 GameObject:
