+++ Assets/Prefabs/KeyLocationClearSphere.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &909444189230491489
+--- !u!1 &7324278031975958964
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,48 +8,46 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 1754046195229052658}
-  - component: {fileID: 6471698742100256899}
-  - component: {fileID: 879477788856803401}
-  - component: {fileID: 8147916920504858516}
-  - component: {fileID: 2514502999786563198}
-  - component: {fileID: 4802554107348960827}
+  - component: {fileID: 7207133435121515787}
+  - component: {fileID: 6064830974049041147}
+  - component: {fileID: 1526781885472236993}
+  - component: {fileID: 9069028906373160953}
   m_Layer: 0
-  m_Name: PaintBall
+  m_Name: KeyLocationClearSphere
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &1754046195229052658
+--- !u!4 &7207133435121515787
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 909444189230491489}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -3.5157695, y: 1.6445246, z: 10.331785}
-  m_LocalScale: {x: 0.2, y: 0.2, z: 0.2}
+  m_GameObject: {fileID: 7324278031975958964}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!33 &6471698742100256899
+--- !u!33 &6064830974049041147
 MeshFilter:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 909444189230491489}
+  m_GameObject: {fileID: 7324278031975958964}
   m_Mesh: {fileID: 10207, guid: 0000000000000000e000000000000000, type: 0}
---- !u!23 &879477788856803401
+--- !u!23 &1526781885472236993
 MeshRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 909444189230491489}
+  m_GameObject: {fileID: 7324278031975958964}
   m_Enabled: 1
   m_CastShadows: 1
   m_ReceiveShadows: 1
@@ -62,7 +60,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: 1b7c38edd9d12684abe4d59acb7c9c2c, type: 2}
+  - {fileID: 2100000, guid: 0d391db818ab34c4eb1855871281f434, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -84,46 +82,20 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!135 &8147916920504858516
-SphereCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 909444189230491489}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 1
-  serializedVersion: 2
-  m_Radius: 0.5
-  m_Center: {x: 0, y: 0, z: 0}
---- !u!54 &2514502999786563198
-Rigidbody:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 909444189230491489}
-  serializedVersion: 2
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_UseGravity: 1
-  m_IsKinematic: 0
-  m_Interpolate: 2
-  m_Constraints: 0
-  m_CollisionDetection: 2
---- !u!114 &4802554107348960827
+--- !u!114 &9069028906373160953
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 909444189230491489}
+  m_GameObject: {fileID: 7324278031975958964}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f390b634016cb7144b24e4944f061147, type: 3}
+  m_Script: {fileID: 11500000, guid: 1a1a5034aa590de43935e3e968f96d0b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  impactSphere: {fileID: 9011068551483308292, guid: 4ac8f342483920a43bc4b1522ffcd39f, type: 3}
-  impactSphereSpawnAmount: 6
+  keyLocationClearSphere: 1
+  upMinValue: 0
+  upMaxValue: 4
+  maxMultiplier: 12
+  lifetime: 5
